// src/app/model/sociedad.ts
export interface Sociedad {
    id?: string;
    nombre: string;
    cif: string;
    direccionSocial: Direccion;
    direccionFiscal: Direccion;
    fechaCreacion?: Date;
    fechaActualizacion?: Date;
}