// src/app/model/oficina.ts
/**
 * Avertia API
 * API definition for Avertia
 *
 * OpenAPI spec version: 1.0.0
 * Contact: info@avertia.es
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BSONObjectID } from './bSONObjectID';
import { Direccion } from './direccion';
import { Perfil } from './perfil';


/**
 * Oficina Object
 */
export interface Oficina {
    id?: string;
    nombre: string;
    direccion: Direccion;
    contacto: Perfil;
    fechaCreacion?: Date;
    fechaActualizacion?: Date;
}

export const keyValuesOficina = [
    {
        value: '@Completa el nombre de la nueva oficina', elements: [
            {
                key: 'nombre', value: '@nombre', formatType: 'string', customClass: 'column-size-1',
                visible: true, editable: true, filtrable: true, customClassInput: 'input-width-4'
            },
        ]
    },
    {
        value: '@Completa el nombre de la nueva oficina', elements: [
            {
                key: 'direccion.direccion', value: '@dirección', customClass: 'column-size-3',
                formatType: 'string', visible: true, editable: true, filtrable: true, customClassInput: 'input-width-4'
            },
            {
                key: 'direccion.cp', value: '@código postal', customClass: 'column-size-1',
                formatType: 'string', visible: false, editable: true, filtrable: true, customClassInput: 'input-width-2'
            },
            {
                key: 'direccion.poblacion', value: '@población', formatType: 'string',
                visible: true, editable: true, filtrable: true, customClassInput: 'input-width-2'
            },
            {
                key: 'direccion.provincia', value: '@provincia', formatType: 'string',
                visible: false, editable: true, filtrable: true, customClassInput: 'input-width-2'
            },
            {
                key: 'direccion.pais', value: '@pais', formatType: 'string',
                visible: true, editable: true, filtrable: true, customClassInput: 'input-width-2'
            },
        ]
    },
    {
        value: '@Completa el nombre de la persona de contacto', elements: [
            {
                key: 'contacto.nombre', value: '@contacto', formatType: 'string',
                visible: true, editable: true, filtrable: true, customClassInput: 'input-width-2'
            },
            {
                key: 'contacto.telefono', value: '@telefono', formatType: 'string',
                visible: true, editable: true, filtrable: true, customClassInput: 'input-width-2'
            }
        ]
    },
];