// src/app/model/direccion.ts
/**
 * Avertia API
 * API definition for Avertia
 *
 * OpenAPI spec version: 1.0.0
 * Contact: info@avertia.es
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Direccion Object
 */
export interface Direccion {
    id: string;
    direccion: string;
    cp: string;
    poblacion: string;
    provincia: string;
    pais: string;
}

export const keyValuesDireccion = [
    { key: 'direccion', value: '@direccion', formatType: 'string', visible: true, editable: true, filtrable: true },
    { key: 'cp',        value: '@cp',        formatType: 'string', visible: true, editable: true, filtrable: true },
    { key: 'poblacion', value: '@poblacion', formatType: 'string', visible: true, editable: true, filtrable: true },
    { key: 'provincia', value: '@provincia', formatType: 'string', visible: true, editable: true, filtrable: true },
    { key: 'pais',      value: '@pais',      formatType: 'string', visible: true, editable: true, filtrable: true },
];