// src/app/model/empleado.ts
export interface Empleado {
    id?: string;
    fechaCreacion?: Date;
    fechaActualizacion?: Date;
    nombre: string;
    apellidos: string;
    direccion?: Direccion;
    email: string;
    telefono?: string;
    sociedad: Sociedad;
    oficina: Oficina;
    ubicacion: string;
    cargos: Array<Rol>;
    salario: string;
    otros?: string;
}