// src/app/model/rol.ts
/**
 * Avertia Security
 * API definition for Avertia Security
 *
 * OpenAPI spec version: 1.0.0
 * Contact: info@avertia.es
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export type Rol = 'Empleado' | 'JefeProyecto' | 'Gerente' | 'Administrativo' | 'Financiero' | 'Socio';

export const Rol = {
    Empleado: 'Empleado' as Rol,
    JefeProyecto: 'JefeProyecto' as Rol,
    Gerente: 'Gerente' as Rol,
    Administrativo: 'Administrativo' as Rol,
    Financiero: 'Financiero' as Rol,
    Socio: 'Socio' as Rol
};